requirements.txt
matplotlib
pyrosbag
numpy
pyrealsense2
Pillow
opencv-python